feat: rename useSession to useAuth
