+++ docs/content/1.getting-started/1.index.md
@@ -56,7 +56,7 @@ Here's the source-code https://github.com/sidebase/nuxt-auth-example of the exam
 ## Application Side Session Management
 
 ::list{type="success"}
-- composable `const { signIn, signOut, status, data, lastRefreshedAt, ... } = useSession()`
+- composable `const { signIn, signOut, status, data, lastRefreshedAt, ... } = useAuth()`
 - Auto-refresh the session periodically
 - Auto-refresh the session on tab-refocus
 - Efficient session fetching: One time on page load, afterwards for specific actions (e.g., on navigation)
