+++ docs/content/1.getting-started/3.quick-start.md
@@ -31,7 +31,7 @@ That's it! You can now use all user-related functionality, for example:
 ::code-group
 ```ts [Application side]
 // file: e.g ~/pages/login.vue
-const { status, data, signIn, signOut } = useSession()
+const { status, data, signIn, signOut } = useAuth()
 
 status.value // Session status: `unauthenticated`, `loading`, `authenticated`
 data.value // Session data, e.g., expiration, user.email, ...
