+++ docs/content/3.application-side/1.index.md
@@ -6,7 +6,7 @@ description: "Application-side usage of `nuxt-auth` for Vue / Nuxt 3 apps."
 
 On the application-side this module offers:
 ::list{type="success"}
-- [`useSession` composable for session access and management](/nuxt-auth/application-side/session-access-and-management)
+- [`useAuth` composable for session access and management](/nuxt-auth/application-side/session-access-and-management)
 - [Creation of custom sign-in pages](/nuxt-auth/application-side/custom-sign-in-page)
 - [Middleware to protect your application on the application side](/nuxt-auth/application-side/protecting-pages)
 ::
