+++ docs/content/3.application-side/2.session-access-and-management.md
@@ -1,6 +1,6 @@
 # Session Access and Management
 
-The `useSession` composable is your main gateway to accessing and manipulating session-state and data. Here's the main methods you can use:
+The `useAuth` composable is your main gateway to accessing and manipulating session-state and data. Here's the main methods you can use:
 
 ```ts
 const {
@@ -12,7 +12,7 @@ const {
   getSession,
   signIn,
   signOut,
-} = useSession()
+} = useAuth()
 
 // Session status, either `unauthenticated`, `loading`, `authenticated`, see https://next-auth.js.org/getting-started/client#signout
 status.value
