+++ docs/content/3.application-side/3.custom-sign-in-page.md
@@ -22,7 +22,7 @@ To create your custom sign-in page you can use `signIn` to directly start a prov
 <script setup lang="ts">
 definePageMeta({ auth: false })
 
-const { signIn } = useSession()
+const { signIn } = useAuth()
 </script>
 ```
 
@@ -68,7 +68,7 @@ If you have not set `enableGlobalAppMiddleware` or have set it to `false` this s
 
 You can handle sign-in errors yourself by calling `signIn` with `redirect: false` and inspecting its result for errors. For example:
 ```ts
-const { signIn } = useSession()
+const { signIn } = useAuth()
 
 const mySignInHandler = async ({ username, password }: { username: string, password: string }) => {
   const { error, url } = await signIn('credentials', { username, password, redirect: false })
