+++ docs/content/3.application-side/4.protecting-pages.md
@@ -87,7 +87,7 @@ Creating a custom middleware is an advanced, experimental option and may result 
 
 To implement your custom middleware:
 1. Create an application-side middleware that applies either globally or is named (see [the Nuxt docs for more](https://nuxt.com/docs/guide/directory-structure/middleware#middleware-directory))
-2. Add logic based on `useSession` to it
+2. Add logic based on `useAuth` to it
 
 When adding the logic, you need to watch out when calling other `async` composable functions. This can lead to `context`-problems in Nuxt, see [the explanation for this here](https://github.com/nuxt/framework/issues/5740#issuecomment-1229197529). In order to avoid these problems, you will need to either:
 - use the undocumented `callWithNuxt` utility when `await`ing other composables,
@@ -98,7 +98,7 @@ Following are examples of both kinds of usage:
 ```ts [direct return]
 // file: ~/middleware/authentication.global.ts
 export default defineNuxtRouteMiddleware((to) => {
-  const { status, signIn } = useSession()
+  const { status, signIn } = useAuth()
 
   // Return immeadiatly if user is already authenticated
   if (status.value === 'authenticated') {
@@ -121,7 +121,7 @@ export default defineNuxtRouteMiddleware((to) => {
   // It's important to do this as early as possible
   const nuxtApp = useNuxtApp()
 
-  const { status, signIn } = useSession()
+  const { status, signIn } = useAuth()
 
   // Return immeadiatly if user is already authenticated
   if (status.value === 'authenticated') {
