+++ docs/content/5.recipes/4.custom-session-data.md
@@ -88,7 +88,7 @@ Notice that in the `jwt` callback we pass both the `access_token` and the `role`
 After that you can access the new fields in the data portion of the session object:
 
 ```ts
-const { status, data } = useSession(); 
+const { status, data } = useAuth();
 
 console.log(data);
 
@@ -103,6 +103,6 @@ output:
 With this approach `data` will automatically be correctly typed.
 
 
-Thanks to 
+Thanks to
 - [@JoaoPedroAS51 from Github](https://github.com/JoaoPedroAS51) for his contribution here: https://github.com/sidebase/nuxt-auth/issues/61#issuecomment-1330747199
 - [@pvlastaridis from Github](https://github.com/pvlastaridis) for his contribution here: https://github.com/sidebase/nuxt-auth/issues/148#issuecomment-1407490921
