+++ src/module.ts
@@ -1,7 +1,7 @@
 import { defineNuxtModule, useLogger, addImportsDir, createResolver, addTemplate, addPlugin, addServerPlugin } from '@nuxt/kit'
 import { defu } from 'defu'
 import { joinURL } from 'ufo'
-import { SupportedProviders } from './runtime/composables/useSession'
+import { SupportedProviders } from './runtime/composables/useAuth'
 
 interface GlobalMiddlewareOptions {
   /**
